Downgrade F2018 requirement to F2003 for NVHPC support (#346)
The Nvidia compilers do not support the assumed rank F2018 feature 
that we use to minimize the code in the `client` and `dataset`. Using
a macro that detects whether the Nvidia compiler is being used,
the declaration is instead assumed-size (`dimension(*)`). The macro
is required as both gfortran and ifort do not see this use of
assumed-size as valid. All tests where library and testers are compiled
with Nvidia 22.3.0 have confirmed to pass. Note: support for this 
should be considered experimental until we add this compiler chain
to our nightly builds.

[ committed by @ashao ]
[ reviewed by @billschereriii ]